+++ src/main/java/com/mudosa/musinsa/settlement/domain/dto/YearlyAggregationDto.java
@@ -13,7 +13,7 @@ public class YearlyAggregationDto {
 
     private Long brandId;
     private Integer year;
-    private Integer totalOrderCount;
+    private Long totalOrderCount;
     private Money totalSalesAmount;
     private Money totalCommissionAmount;
     private Money totalTaxAmount;
@@ -30,7 +30,7 @@ public class YearlyAggregationDto {
     ) {
         this.brandId = brandId;
         this.year = year;
-        this.totalOrderCount = totalOrderCount.intValue();
+        this.totalOrderCount = totalOrderCount;
         this.totalSalesAmount = new Money(totalSalesAmount);
         this.totalCommissionAmount = new Money(totalCommissionAmount);
         this.totalTaxAmount = new Money(totalTaxAmount);
